msm: camera: Fix for enabling front camera
Adding int type in comparison as csiphy_clk is defined
as unsigned int whereas it is being assigned -1.

+++ drivers/media/platform/msm/camera_v2/sensor/csiphy/msm_csiphy.c
@@ -78,7 +78,7 @@ static int msm_csiphy_lane_config(struct csiphy_device *csiphy_dev,
 		return -EINVAL;
 	}
 
-	clk_rate = (csiphy_params->csiphy_clk > 0)
+	clk_rate = ((int)csiphy_params->csiphy_clk > 0)
 			? csiphy_params->csiphy_clk :
 			csiphy_dev->csiphy_max_clk;
 	round_rate = clk_round_rate(
